requirements.txt
alembic~=1.8
asyncpg~=0.26
fastapi[all]~=0.85
passlib[bcrypt]
psycopg2-binary~=2.9
python-dotenv~=0.21
python-jose[cryptography]
SQLAlchemy~=1.4
